Consolidated Cooldown Graphs into one file
