--- src/analysis/retail/priest/shadow/modules/guide/LongCooldownGraphSubSection.tsx
@@ -1,63 +1,0 @@
-import { Talent } from 'common/TALENTS/types';
-import { SubSection, useAnalyzer, useInfo } from 'interface/guide';
-import TALENTS from 'common/TALENTS/priest';
-//import SPELLS from 'common/SPELLS/priest';
-//import { SpellLink } from 'interface';
-import CastEfficiency from 'parser/shared/modules/CastEfficiency';
-import CastEfficiencyBar from 'parser/ui/CastEfficiencyBar';
-import { GapHighlight } from 'parser/ui/CooldownBar';
-import { Trans } from '@lingui/macro';
-
-type Cooldown = {
-  talent: Talent;
-  extraTalents?: Talent[];
-};
-
-const cooldownsToCheck: Cooldown[] = [
-  { talent: TALENTS.POWER_INFUSION_TALENT },
-  { talent: TALENTS.DARK_ASCENSION_TALENT },
-  { talent: TALENTS.VOID_ERUPTION_TALENT },
-  { talent: TALENTS.MINDBENDER_SHADOW_TALENT },
-];
-
-const LongCooldownGraphSubsection = () => {
-  const info = useInfo();
-  const castEfficiency = useAnalyzer(CastEfficiency);
-  if (!info || !castEfficiency) {
-    return null;
-  }
-
-  const cooldowns = cooldownsToCheck.filter((cooldown) => {
-    const hasTalent = info.combatant.hasTalent(cooldown.talent);
-    const hasExtraTalents =
-      cooldown.extraTalents?.reduce(
-        (acc, talent) => acc && info.combatant.hasTalent(talent),
-        true,
-      ) ?? true;
-    return hasTalent && hasExtraTalents;
-  });
-  const hasTooManyCasts = cooldowns.some((cooldown) => {
-    const casts = castEfficiency.getCastEfficiencyForSpell(cooldown.talent)?.casts ?? 0;
-    return casts >= 10;
-  });
-
-  return (
-    <SubSection>
-      <Trans id="guide.priest.shadow.sections.cooldowns.graph">
-        <strong>Major Cooldown Graph</strong> - this graph shows when you used your cooldowns and
-        how long you waited to use them again. You should use these cooldowns together to maximize
-        the damage they can deal.
-      </Trans>
-      {cooldowns.map((cooldownCheck) => (
-        <CastEfficiencyBar
-          key={cooldownCheck.talent.id}
-          spellId={cooldownCheck.talent.id}
-          gapHighlightMode={GapHighlight.FullCooldown}
-          minimizeIcons={hasTooManyCasts}
-        />
-      ))}
-    </SubSection>
-  );
-};
-
-export default LongCooldownGraphSubsection;
